+++ lib/onlyoffice_documentserver_testing_framework/test_instance_docs/integration_example_api.rb
@@ -16,17 +16,32 @@ module OnlyofficeDocumentserverTestingFramework
       http_get("#{@api_endpoint}/files/")
     end
 
+    # Upload file by it's path
+    # @param [String] file_path
+    # @return [String] name of uploaded file
+    def upload_file(file_path)
+      url = URI("#{@api_endpoint}/upload/")
+
+      request = Net::HTTP::Post.new(url)
+      form_data = [['uploadedFile', File.open(file_path)]]
+      request.set_form(form_data, 'multipart/form-data')
+      response = http_from_url(url).request(request)
+      JSON.parse(response.read_body)['filename']
+    end
+
     private
 
-    def http_get(url)
+    def http_from_url(url)
       uri = URI(url)
 
-      https = Net::HTTP.new(uri.host, uri.port)
-      https.use_ssl = true if uri.scheme == 'https'
+      http = Net::HTTP.new(uri.host, uri.port)
+      http.use_ssl = true if uri.scheme == 'https'
+      http
+    end
 
-      request = Net::HTTP::Get.new(uri)
-
-      response = https.request(request)
+    def http_get(url)
+      request = Net::HTTP::Get.new(url)
+      response = http_from_url(url).request(request)
       JSON.parse(response.read_body)
     end
   end