+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 * Add several new errors to ignored list
 * Freeze dev dependencies version in `Gemfile.lock`
-* Store `DocTestSiteFunctions.supported_languages` as class method, 
+* Store `DocTestSiteFunctions.supported_languages` as class method,
   not constant
 
 ## 0.2.0 (2020-09-13)