+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change log
 
+## 0.3.0 (2021-03-24)
+
 ## master (unreleased)
 
 ### New Features